add arbitrary size limit to PreviousSentData
The current implementation for checking duplicate data is realistically unsustainable because eventually the PreviousSentData variable will become so big that iterating through it will cause enormous lags and eventually leading to server unresponsiveness.

This PR simply adds an arbitrary size check and clears the PreviousSentData when reached. However, I am not certain that the size is enough to cover potential duplicate data, so obviously this is up for discussion. I believe we should keep the PreviousSentData size as low as possible to ensure proper server performance.