server-side password validation on web socket connection
This will prevent a modified/patched SIT client from accessing a password-protected server if the client-side password validation is removed. The end result is not elegant but effective: the connection will be closed while joining the raid and you will be stuck in "Waiting for 1 player..." loading screen if the password provided does not match the server's. However, a legit SIT client will never end up in this scenario because the password validation is handled properly in the SIT's JoinMatch function prior to the web socket connection.