support all bundle file extension
This issue was brought up by SiulSC on Discord.

If bundle files have a different extension than .bundle, SPT will return an UNHANDLED exception and won't download the bundle correctly. However, custom bundle extensions are managed on normal standard SPT, so not sure why we're getting this issue with SITCoop. This PR aims to fix this issue.

Tested using this mod: https://hub.sp-tarkov.com/files/file/889-goblin-king/ which is using .goblin with some of its bundles.
Confirmed to be working with this PR.