login password requirement + override refactoring
This PR makes it so you now require a password to log in in order to receive a session ID. Most account operations such as profile wipe, change password, etc. requires a successful login, so this will prevent any malicious operations from succeeding. After raid profile save requires a session ID which can only be obtained if you log in (with a password).

Launchers will now need to be able to process the "INVALID_PASSWORD" return value when querying /launcher/profile/login for instances where the password provided is incorrect or missing.

The overrides now have been refactored in their own classes for code clarity and simplicity.